requirements.txt
python-dotenv
openai
python-bithumb>=0.1.2
schedule
streamlit
plotly
requests